SDTM code lists

sdtm_code_lists is a subset of the data provided in provided in sdtm_terminology. It contains only the terms for code lists from CDISC SDTM Terminology.

Usage

sdtm_code_lists

Format

A tibble providing a data set of controlled terms for SDTM code lists.

code

A C-code with a leading character C followed by a sequence of digits used for uniquely identifying each concept in NCI Thesaurus (NCIt), including all CDISC concepts.

term

This is the specific value expected for submissions.

is_extensible

Defines if controlled terms may be added to the codelist. New terms may be added to existing codelist values when the codelist has value TRUE as long as the terms are not duplicates or synonyms of existing terms. The expectation is that sponsors will use the published controlled terminology as a standard baseline and codelists defined as extensible may have terms added by the sponsor internally. For codelists with FALSE sponsor organizations may not add to the published lists. Suggestions for new terms may be requested through the New Term Request Page.

cl_name

Contains the descriptive name of the codelist.

synonym

This identifies the applicable synonyms for the CDISC submission value indicated in column term. These terms should not be submitted, but are included for collection and mapping purposes. Note: this is especially important in instances where a test name or parameter test name contains a corresponding test code or parameter test code.

definition

This identifies the value of an ALT_DEFINITION property of a CDISC source for a particular concept. In many cases an existing NCI definition has been used. The source for a definition is noted in parentheses (e.g. NCI, CDISC glossary).

nci_term

This identifies the value of a PREFERRED_NAME property for a concept in NCIt. This column designates the human readable, fully specified preferred term corresponding to the NCI C-code, and is especially helpful for searching NCIt to get the entire concept with links to all instances of the term.

Source

https://evs.nci.nih.gov/ftp1/CDISC/SDTM/

See also

Examples

sdtm_code_lists
#> # A tibble: 1,081 × 7
#>    code    term          is_extensible cl_name       synonym definition nci_term
#>    <chr>   <chr>         <lgl>         <chr>         <chr>   <chr>      <chr>   
#>  1 C141657 TENMW1TC      FALSE         10-Meter Wal… 10-Met… 10-Meter … CDISC F…
#>  2 C141656 TENMW1TN      FALSE         10-Meter Wal… 10-Met… 10-Meter … CDISC F…
#>  3 C141663 A4STR1TC      FALSE         4-Stair Asce… 4-Stai… 4-Stair A… CDISC F…
#>  4 C141662 A4STR1TN      FALSE         4-Stair Asce… 4-Stai… 4-Stair A… CDISC F…
#>  5 C141661 D4STR1TC      FALSE         4-Stair Desc… 4-Stai… 4-Stair D… CDISC F…
#>  6 C141660 D4STR1TN      FALSE         4-Stair Desc… 4-Stai… 4-Stair D… CDISC F…
#>  7 C115388 SIXMW1TC      FALSE         6 Minute Wal… 6 Minu… 6 Minute … CDISC F…
#>  8 C115387 SIXMW1TN      FALSE         6 Minute Wal… 6 Minu… 6 Minute … CDISC F…
#>  9 C182464 AIMS0101T07OR FALSE         Abnormal Inv… Abnorm… Abnormal … CDISC C…
#> 10 C182465 AIMS0108T09OR FALSE         Abnormal Inv… Abnorm… Abnormal … CDISC C…
#> # ℹ 1,071 more rows