SDTM Terminology

sdtm_terminology provides CDISC SDTM Terminology data in tibble format. It follows CDISC original structure with a few modifications, see Details section.

Usage

sdtm_terminology

Format

A tibble providing a data set of controlled terms: either terms for code lists or individual terms representing CDISC concepts subjected to controlled terminology. These two cases can be distinguished by the variable cl_code: the parent code list code. A term representing a code list won't have a parent code list code, i.e. cl_code is NA.

code

A C-code with a leading character C followed by a sequence of digits used for uniquely identifying each concept in NCI Thesaurus (NCIt), including all CDISC concepts.

term

This is the specific value expected for submissions.

cl_code

An NCIt C-code assigned to the parent codelist name. This code is repeated for each concept (aka permissible value) belonging to a codelist.

is_extensible

Defines if controlled terms may be added to the codelist. New terms may be added to existing codelist values when the codelist has value TRUE as long as the terms are not duplicates or synonyms of existing terms. The expectation is that sponsors will use the published controlled terminology as a standard baseline and codelists defined as extensible may have terms added by the sponsor internally. For codelists with FALSE sponsor organizations may not add to the published lists. Suggestions for new terms may be requested through the New Term Request Page.

cl_name

Contains the descriptive name of the codelist.

synonym

This identifies the applicable synonyms for the CDISC submission value indicated in column term. These terms should not be submitted, but are included for collection and mapping purposes. Note: this is especially important in instances where a test name or parameter test name contains a corresponding test code or parameter test code.

definition

This identifies the value of an ALT_DEFINITION property of a CDISC source for a particular concept. In many cases an existing NCI definition has been used. The source for a definition is noted in parentheses (e.g. NCI, CDISC glossary).

nci_term

This identifies the value of a PREFERRED_NAME property for a concept in NCIt. This column designates the human readable, fully specified preferred term corresponding to the NCI C-code, and is especially helpful for searching NCIt to get the entire concept with links to all instances of the term.

Source

https://evs.nci.nih.gov/ftp1/CDISC/SDTM/

Details

Column names have been shortened and changed to follow snake case style.

The original variable Codelist Extensible (Yes/No) whose type is text, is here renamed to is_extensible and made a logical vector with values recoded from "No" and "Yes" to FALSE and TRUE, respectively.

See also

Examples

sdtm_terminology
#> # A tibble: 40,816 × 8
#>    code    term        cl_code is_extensible cl_name synonym definition nci_term
#>    <chr>   <chr>       <chr>   <lgl>         <chr>   <chr>   <chr>      <chr>   
#>  1 C141657 TENMW1TC    NA      FALSE         10-Met… 10-Met… 10-Meter … CDISC F…
#>  2 C174106 TENMW101    C141657 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  3 C141700 TENMW102    C141657 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  4 C147592 TENMW103    C141657 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  5 C141701 TENMW104    C141657 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  6 C141656 TENMW1TN    NA      FALSE         10-Met… 10-Met… 10-Meter … CDISC F…
#>  7 C141701 TENMW1-Tes… C141656 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  8 C141700 TENMW1-Tim… C141656 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#>  9 C174106 TENMW1-Was… C141656 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#> 10 C147592 TENMW1-Wea… C141656 NA            10-Met… TENMW1… 10-Meter … 10-Mete…
#> # ℹ 40,806 more rows